refactoring: Moved ScreenSize.kt to the domain/constants
